app/requirements.txt
fastapi
uvicorn
python-telegram-bot==22.2
python-dotenv==1.1.1
loguru==0.7.3
slowapi==0.1.9
